+++ src/FsPickler.CSharp/Properties/AssemblyInfo.cs
@@ -3,11 +3,13 @@ using System.Reflection;
 
 [assembly: AssemblyProductAttribute("FsPickler")]
 [assembly: AssemblyCopyrightAttribute("© Eirik Tsarpalis.")]
-[assembly: AssemblyVersionAttribute("3.2.0")]
-[assembly: AssemblyFileVersionAttribute("3.2.0")]
+[assembly: AssemblyVersionAttribute("4.0.0")]
+[assembly: AssemblyFileVersionAttribute("4.0.0")]
 namespace System {
     internal static class AssemblyVersionInformation {
-        internal const string Version = "3.2.0";
-        internal const string InformationalVersion = "3.2.0";
+        internal const System.String AssemblyProduct = "FsPickler";
+        internal const System.String AssemblyCopyright = "© Eirik Tsarpalis.";
+        internal const System.String AssemblyVersion = "4.0.0";
+        internal const System.String AssemblyFileVersion = "4.0.0";
     }
 }