+++ src/FsPickler.Json/JsonReader.fs
@@ -49,9 +49,9 @@ type internal JsonPickleReader (jsonReader : JsonReader, omitHeader, useCustomSe
             else
                 do jsonReader.MoveNext()
                 let version = jsonReader.ReadPrimitiveAs<string> false "FsPickler"
-                if version <> formatv2000 then
+                if version <> formatv4000 then
                     let v = Version(version)
-                    if version = formatv0960 || version = formatv1200 || version = formatv1400 then
+                    if version = formatv0960 || version = formatv1200 || version = formatv1400 || version = formatv2000 then
                         raise <| new FormatException(sprintf "JSON format version %O no longer supported." v)
                     else
                         raise <| new FormatException(sprintf "Unrecognized JSON format version %O." v)
@@ -181,9 +181,7 @@ type internal JsonPickleReader (jsonReader : JsonReader, omitHeader, useCustomSe
         member __.ReadChar tag = let value = jsonReader.ReadPrimitiveAs<string> (omitTag ()) tag in value.[0]
         member __.ReadString tag = jsonReader.ReadPrimitiveAs<string> (omitTag ()) tag
 
-#if !NET35
         member __.ReadBigInteger tag = jsonReader.ReadPrimitiveAs<string> (omitTag ()) tag |> BigInteger.Parse
-#endif
 
         member __.ReadGuid tag = 
             if isBsonReader then