+++ src/FsPickler/PicklerGeneration/PicklerGenerator.fs
@@ -10,7 +10,6 @@ open System.Reflection
 open System.Runtime.Serialization
 
 open TypeShape
-open TypeShape_ISerializableExtensions
 
 open MBrace.FsPickler
 open MBrace.FsPickler.Reflection
@@ -26,7 +25,7 @@ type PicklerGenerator =
         with UnsupportedShape t -> raise <| NonSerializableTypeException(t)
         
     /// Constructs a pickler for a given shape
-    static member Create (resolver : IPicklerResolver) (shape : TypeShape) : Pickler =
+    static member Create (registry : ICustomPicklerRegistry) (resolver : IPicklerResolver) (shape : TypeShape) : Pickler =
         let isUnsupportedType (t:Type) =
             t.IsPointer 
             || t = typeof<System.Reflection.Pointer>
@@ -34,6 +33,8 @@ type PicklerGenerator =
             || t.IsCOMObject
             || t.IsImport
             || t.IsMarshalByRef
+
+        let (|PicklerFactory|_|) (shape : TypeShape) = registry.TryGetPicklerFactory shape.Type
 
         match shape with
         | _ when isUnsupportedType shape.Type -> raise <| NonSerializableTypeException shape.Type
@@ -57,20 +58,18 @@ type PicklerGenerator =
         | Shape.TimeSpan -> new TimeSpanPickler() :> _
         | Shape.ByteArray -> ArrayPickler.CreateByteArrayPickler() :> _
         | Shape.Unit -> new UnitPickler() :> _
-#if !NET35
         | :? TypeShape<bigint> -> new BigIntPickler() :> _
-#endif
         | :? TypeShape<System.Object> -> CompositePickler.ObjectPickler :> _
         | :? TypeShape<AssemblyInfo> -> ReflectionPicklers.CreateAssemblyInfoPickler() :> _
         | :? TypeShape<AssemblyName> -> ReflectionPicklers.CreateAssemblyNamePickler resolver :> _
         | :? TypeShape<Assembly> -> ReflectionPicklers.CreateAssemblyPickler resolver :> _
         | :? TypeShape<MemberInfo> -> ReflectionPicklers.CreateMemberInfoPickler resolver :> _
         | :? TypeShape<System.DBNull> -> new DBNullPickler() :> _
-        | _ when PicklerPluginRegistry.ContainsFactory shape.Type ->
-            shape.Accept {
-                new ITypeShapeVisitor<Pickler> with
-                    member __.Visit<'T> () = PicklerPluginRegistry.GetPicklerFactory<'T>() resolver :> Pickler 
-            }
+        | PicklerFactory factory -> 
+            let pickler = factory resolver
+            if pickler.Type <> shape.Type then
+                raise <| PicklerGenerationException(shape.Type, "unexpected pickler type from custom pickler generator.")
+            pickler
 
         | Shape.Nullable s ->
             s.Accept {
@@ -82,25 +81,13 @@ type PicklerGenerator =
         | Shape.Array s ->
             s.Accept {
                 new IArrayVisitor<Pickler> with
-                    member __.Visit<'T> () = ArrayPickler.Create<'T>(resolver) :> _
-            }
-
-        | Shape.Array2D s ->
-            s.Accept {
-                new IArray2DVisitor<Pickler> with
-                    member __.Visit<'T> () = ArrayPickler.Create2D<'T>(resolver) :> _
-            }
-
-        | Shape.Array3D s ->
-            s.Accept {
-                new IArray3DVisitor<Pickler> with
-                    member __.Visit<'T> () = ArrayPickler.Create3D<'T>(resolver) :> _
-            }
-
-        | Shape.Array4D s ->
-            s.Accept {
-                new IArray4DVisitor<Pickler> with
-                    member __.Visit<'T> () = ArrayPickler.Create4D<'T>(resolver) :> _
+                    member __.Visit<'T> rank = 
+                        match rank with
+                        | 1 -> ArrayPickler.Create<'T>(resolver) :> _
+                        | 2 -> ArrayPickler.Create2D<'T>(resolver) :> _
+                        | 3 -> ArrayPickler.Create3D<'T>(resolver) :> _
+                        | 4 -> ArrayPickler.Create4D<'T>(resolver) :> _
+                        | _ -> raise <| NonSerializableTypeException(shape.Type, "Array ranks more than 4 are not supported.")
             }
 
         | Shape.Tuple1 s ->
@@ -243,7 +230,12 @@ type PicklerGenerator =
             // NB: DataContractAttribute for enum types handled by enum pickler
             s.Accept { 
                 new IEnumVisitor<Pickler> with
-                    member __.Visit<'E, 'U when 'E : enum<'U>>() = EnumPickler.Create<'E, 'U>(resolver) :> _
+                    member __.Visit<'E, 'U when 'E : enum<'U> 
+                                            and 'E : struct 
+                                            and 'E :> ValueType
+                                            and 'E : (new : unit -> 'E)> () = 
+
+                        EnumPickler.Create<'E, 'U>(resolver) :> _
             }
 
         | _ when containsAttr<DataContractAttribute> shape.Type ->
@@ -262,13 +254,13 @@ type PicklerGenerator =
         | Shape.FSharpUnion _ as s ->
             s.Accept {
                 new ITypeShapeVisitor<Pickler> with
-                    member __.Visit<'U> () = FsUnionPickler.Create<'U>(resolver) :> _
+                    member __.Visit<'U> () = FsUnionPickler.Create<'U> registry resolver :> _
             }
 
         | Shape.FSharpRecord _ as s ->
             s.Accept {
                 new ITypeShapeVisitor<Pickler> with
-                    member __.Visit<'R> () = FsRecordPickler.Create<'R>(resolver) :> _
+                    member __.Visit<'R> () = FsRecordPickler.Create<'R> registry resolver :> _
             }
 
         | shape when shape.Type.IsAbstract ->
@@ -280,33 +272,33 @@ type PicklerGenerator =
         | Shape.Exception s ->
             s.Accept {
                 new IExceptionVisitor<Pickler> with
-                    member __.Visit<'exn when 'exn :> exn>() = 
-                        if s.IsFSharpException then FsExceptionPickler.Create<'exn>(resolver) :> _
+                    member __.Visit<'exn when 'exn :> exn and 'exn : not struct and 'exn : null>() = 
+                        if s.IsFSharpException then FsExceptionPickler.Create<'exn> registry resolver :> _
                         else
                             match tryGetISerializableCtor typeof<'exn> with
-                            | None -> ISerializablePickler.CreateNonISerializableExceptionPickler<'exn>(resolver) :> _
+                            | None -> ISerializablePickler.CreateNonISerializableExceptionPickler<'exn> registry resolver :> _
                             | Some _ -> ISerializablePickler.Create<'exn>() :> _
             }
 
         | Shape.ISerializable s ->
             s.Accept {
                 new ISerializableVisitor<Pickler> with
-                    member __.Visit<'T when 'T :> ISerializable> () =
+                    member __.Visit<'T when 'T :> ISerializable> (ss:ShapeISerializable<'T>): Pickler =
                         match tryGetISerializableCtor typeof<'T> with
                         | Some _ -> ISerializablePickler.Create<'T>() :> _
                         | None -> ISerializablePickler.CreateObjectReferencePickler<'T>() :> _
             }
 
-        | shape when shape.Type.IsValueType ->
-            shape.Accept {
-                new ITypeShapeVisitor<Pickler> with
-                    member __.Visit<'T>() = StructFieldPickler.Create<'T>(resolver) :> Pickler
+        | Shape.Struct s ->
+            s.Accept {
+                new IStructVisitor<Pickler> with
+                    member __.Visit<'T when 'T : struct>() = StructFieldPickler.Create<'T>(resolver) :> Pickler
             }
 
-        | _ ->
-            shape.Accept {
-                new ITypeShapeVisitor<Pickler> with
-                    member __.Visit<'T>() = ClassFieldPickler.Create<'T>(resolver) :> Pickler
+        | Shape.NotStruct s ->
+            s.Accept {
+                new INotStructVisitor<Pickler> with
+                    member __.Visit<'T when 'T : not struct and 'T : null>() = ClassFieldPickler.Create<'T> registry resolver :> Pickler
             }
 
     /// Constructs a blank, uninitialized pickler object