+++ src/FsPickler/PicklerGeneration/PicklerResolution.fs
@@ -14,7 +14,7 @@ let isSerializable (result : Exn<Pickler>) =
     | Error _ -> false
 
 /// reflection - based pickler resolution
-let resolvePickler (resolver : IPicklerResolver) 
+let resolvePickler (registry : ICustomPicklerRegistry) (resolver : IPicklerResolver) 
                     (mkEarlyBinding : Pickler -> unit) 
                     (isPicklerReferenced : Pickler -> bool) (t : Type) =
 
@@ -67,7 +67,7 @@ let resolvePickler (resolver : IPicklerResolver)
         let p =
             match result with
             | Some p -> p
-            | None -> PicklerGenerator.Create resolver shape
+            | None -> PicklerGenerator.Create registry resolver shape
 
         // step 5: pickler generation complete, copy data to uninitialized binding
         if isPicklerReferenced p0 then CompositePickler.Copy(p, p0) ; Success p0
@@ -90,7 +90,7 @@ type private GenerationState =
 
 /// recursively generates picklers required for given type, 
 /// storing results in global cache when completed.
-let generatePickler (globalCache : ICache<Type, Exn<Pickler>>) (t : Type) =
+let generatePickler (registry : ICustomPicklerRegistry) (globalCache : ICache<Type, Exn<Pickler>>) (t : Type) =
     // a temporary local cache is used to store generated picklers.
     // this serves the purpose of providing recursive bindings rectypes.
     // a local cache keeps the global cache from being contaminated with partial state.
@@ -117,8 +117,10 @@ let generatePickler (globalCache : ICache<Type, Exn<Pickler>>) (t : Type) =
                 p
             | Some (_,p) -> p
             | None ->
-                let p = resolvePickler resolver (fun p -> localCache.Add(p.Type, (GenerationState.Fresh, Success p))) 
-                                                (fun p -> let s,_ = localCache.[p.Type] in s = GenerationState.Referenced) t
+                let p = resolvePickler registry resolver 
+                                (fun p -> localCache.Add(p.Type, (GenerationState.Fresh, Success p))) 
+                                (fun p -> let s,_ = localCache.[p.Type] in s = GenerationState.Referenced) t
+
                 localCache.[t] <- (GenerationState.Completed, p)
                 p