+++ src/FsPickler/Utils/Reflection.fs
@@ -1,9 +1,10 @@
 ﻿module internal MBrace.FsPickler.Reflection
 
 open System
-open System.Text.RegularExpressions
+open System.Collections.Generic
 open System.Reflection
 open System.Runtime.Serialization
+open System.Text.RegularExpressions
 
 open Microsoft.FSharp.Reflection
 
@@ -120,16 +121,14 @@ let isExceptionDispatchInfo (t : Type) =
 /// walks up the type hierarchy, gathering all instance members
 let gatherMembers (t : Type) =
     // resolve conflicts, index by declaring type and field name
-    let gathered = ref Map.empty<string * string, (* index *) int * MemberInfo>
-    let i = ref 0
+    let gathered = new Dictionary<string * string, MemberInfo>()
 
     let rec gather (t : Type) =
         let members = t.GetMembers(allFields)
         for m in members do
             let k = m.DeclaringType.AssemblyQualifiedName, m.ToString()
-            if not <| gathered.Value.ContainsKey k then
-                gathered := gathered.Value.Add(k, (!i, m))
-                incr i
+            if not <| gathered.ContainsKey k then
+                gathered.Add(k, m)
 
         match t.BaseType with
         | null -> ()
@@ -138,11 +137,11 @@ let gatherMembers (t : Type) =
 
     do gather t
 
-    gathered.Value 
-    |> Map.toSeq
-    |> Seq.map snd
-    |> Seq.sortBy fst // sort by index; this is to preserve member serialization ordering
-    |> Seq.map snd
+    gathered
+    // sort by name since Type.GetMembers() is non deterministic an unordered. https://github.com/mbraceproject/FsPickler/issues/92
+    // Since names are not unique in class hierarchies, we use the full key here being (assemblyQualifiedName * FieldName).
+    |> Seq.sortBy (fun kv -> fst kv.Key, kv.Value.Name)
+    |> Seq.map (fun kv -> kv.Value)
     |> Seq.toArray
 
 let gatherSerializedFields (t : Type) =