+++ tests/FsPickler.Tests/SerializerTests.fs
@@ -140,11 +140,8 @@ type ``FsPickler Serializer Tests`` (format : string) as self =
     [<Test; Category("Primitives")>]
     member __.``1. Primitive: System.Guid`` () = Check.QuickThrowOnFail<Guid> testEquals
 
-#if NET35
-#else
     [<Test; Category("Primitives")>]
     member __.``1. Primitive: bigint`` () = Check.QuickThrowOnFail<bigint> testEquals
-#endif
 
     [<Test; Category("Bytes")>]
     member __.``1. Primitive: byte []`` () = testEquals (null : byte []) ; Check.QuickThrowOnFail<byte []> testEquals
@@ -270,12 +267,21 @@ type ``FsPickler Serializer Tests`` (format : string) as self =
         let bs = [|1uy .. 100uy|]
         let ys,zs = testRoundtrip(bs,bs)
         obj.ReferenceEquals(ys,zs) |> should equal refEq
-
+        
     [<Test; Category("Generic BCL Types")>]
     member __.``3. Array: int * string`` () = __.CheckArray<int * string> ()
 
     [<Test; Category("Generic BCL Types")>]
     member __.``3. Array: string * (int * decimal)`` () = __.CheckArray<int * string> ()
+
+    [<Test; Category("Generic BCL Types")>]
+    member __.``3. Array: struct(int * string)`` () = __.CheckArray<struct(int * string)> ()
+    
+    [<Test; Category("Generic BCL Types")>]
+    member __.``3. Array: struct(string * (int * decimal))`` () = __.CheckArray<struct(string * (int * decimal))> ()
+    
+    [<Test; Category("Generic BCL Types")>]
+    member __.``3. Array: struct(string * struct(int * decimal))`` () = __.CheckArray<struct(string * struct (int * decimal))> ()
 
 
     //
@@ -297,11 +303,43 @@ type ``FsPickler Serializer Tests`` (format : string) as self =
         Check.QuickThrowOnFail<string * byte * TimeSpan * Guid * int * uint64 * decimal * int * int> testEquals
 
     [<Test; Category("Generic BCL Types")>]
+    member __.``4. BCL: struct tuple simple`` () =
+        Check.QuickThrowOnFail<ValueTuple<string>> testEquals
+        Check.QuickThrowOnFail<struct(string * byte)> testEquals
+        Check.QuickThrowOnFail<struct(string * byte * TimeSpan)> testEquals
+        Check.QuickThrowOnFail<struct(string * byte * TimeSpan * Guid)> testEquals
+        Check.QuickThrowOnFail<struct(string * byte * TimeSpan * Guid * int)> testEquals
+        Check.QuickThrowOnFail<struct(string * byte * TimeSpan * Guid * int * uint64)> testEquals
+        Check.QuickThrowOnFail<struct(string * byte * TimeSpan * Guid * int * uint64 * decimal)> testEquals
+        Check.QuickThrowOnFail<struct(string * byte * TimeSpan * Guid * int * uint64 * decimal * int)> testEquals
+        Check.QuickThrowOnFail<struct(string * byte * TimeSpan * Guid * int * uint64 * decimal * int * int)> testEquals
+
+    [<Test; Category("Generic BCL Types")>]
     member __.``4. BCL: tuple nested`` () =
         Check.QuickThrowOnFail<int * (string * decimal)> testEquals
         Check.QuickThrowOnFail<(int * (bool * string)) * (string * int16)> testEquals
         Check.QuickThrowOnFail<(int * (bool * (sbyte * string * uint32) * (string * string)))> testEquals
 
+        
+    [<Test; Category("Generic BCL Types")>]
+    member __.``4. BCL: struct tuple nested`` () =
+        Check.QuickThrowOnFail<struct(int * struct (string * decimal))> testEquals
+        Check.QuickThrowOnFail<struct ((int * struct (bool * string)) * struct (string * int16))> testEquals
+        Check.QuickThrowOnFail<struct (int * struct (bool * struct (sbyte * string * uint32) * struct (string * string)))> testEquals
+
+    
+    [<Test; Category("Generic BCL Types")>]
+    member __.``4. BCL: tuple/struct tuple mixed`` () =
+        Check.QuickThrowOnFail<int * struct(string * decimal)> testEquals
+        Check.QuickThrowOnFail<struct (int * (string * decimal))> testEquals
+        Check.QuickThrowOnFail<struct (int * (bool * string)) * (string * int16)> testEquals
+        Check.QuickThrowOnFail<struct (int * (bool * string)) * struct (string * int16)> testEquals
+        Check.QuickThrowOnFail<(int * struct (bool * string)) * (string * int16)> testEquals
+        Check.QuickThrowOnFail<(int * (bool * (sbyte * string * uint32) * struct (string * string)))> testEquals
+        Check.QuickThrowOnFail<struct (int * (bool * (sbyte * string * uint32) * (string * string)))> testEquals
+        Check.QuickThrowOnFail<(int * struct (bool * (sbyte * string * uint32) * (string * string)))> testEquals
+        Check.QuickThrowOnFail<(int * (bool * struct (sbyte * string * uint32) * (string * string)))> testEquals
+        Check.QuickThrowOnFail<(int * (bool * struct (sbyte * string * uint32) * struct (string * string)))> testEquals
 
     // exceptions
 
@@ -369,6 +407,8 @@ type ``FsPickler Serializer Tests`` (format : string) as self =
 
     [<Test; Category("Generic BCL Types")>]
     member __.``4. BCL: System.Collections.Generic.HashSet`` () =
+        if runsOnMono then () else // skip due to mono 5.2 bug
+
         let testSet (data : seq<'T>) =
             let data = data |> Seq.distinct |> Seq.toList
             let d = new HashSet<'T>(data)
@@ -536,11 +576,17 @@ type ``FsPickler Serializer Tests`` (format : string) as self =
     member __.``5. Object: large pair sequence`` () =
         let pairs = seq { for i in 1 .. 1000000 -> string i,i }
         __.TestSequenceRoundtrip pairs
-
+        
     [<Test; Category("FsPickler Generic tests")>]
     member __.``5. Object: record sequence`` () =
         let records = seq { for i in 1 .. 10000 -> { Int = i ; String = string i ; Tuple = (i, "const") } }
         __.TestSequenceRoundtrip records
+
+    [<Test; Category("FsPickler Generic tests")>]
+    member __.``5. Object: struct record sequence`` () =
+        let records = seq { for i in 1 .. 10000 -> { SInt = i ; SString = string i ; STuple = (i, "const") } }
+        __.TestSequenceRoundtrip records
+
 
     [<Test; Category("FsPickler Generic tests")>]
     member __.``5. Object: rec sequence`` () =
@@ -731,8 +777,6 @@ type ``FsPickler Serializer Tests`` (format : string) as self =
     [<Test ; Category("Custom types")>] 
     member __.``6. Custom: Generic ISerializable class`` () = testEquals <| GenericISerializableClass<int * string>(42, "fortyTwo", (42, "fortyTwo"))
 
-#if NET35
-#else
     [<Test ; Category("Custom types")>] 
     member __.``6. Custom: ISerializable class with IObjectReference template`` () = 
         let x = new ObjRef1(0)
@@ -762,7 +806,6 @@ type ``FsPickler Serializer Tests`` (format : string) as self =
         let x = new DataContractObjRef(0)
         let x' = testRoundtrip x
         x'.Value |> should equal 42
-#endif
 
     [<Test ; Category("Custom types")>] 
     member __.``6. Custom: Data Contract class`` () =
@@ -877,6 +920,10 @@ type ``FsPickler Serializer Tests`` (format : string) as self =
     [<Test; Category("FSharp type tests")>]
     member __.``7. FSharp: simple union`` () =
         Check.QuickThrowOnFail<SimpleDU> testEquals
+        
+    [<Test; Category("FSharp type tests")>]
+    member __.``7. FSharp: struct union`` () =
+        Check.QuickThrowOnFail<StructDU> testEquals
 
     [<Test; Category("FSharp type tests")>]
     member __.``7. FSharp: recursive union`` () =
@@ -915,10 +962,14 @@ type ``FsPickler Serializer Tests`` (format : string) as self =
 
         serializer.UnPickle(data, pickler = (getTreeForestPicklers Pickler.int |> fst))
         |> should equal (nTree 6)
-
+        
     [<Test; Category("FSharp type tests")>]
     member __.``7. FSharp: record`` () = 
         Check.QuickThrowOnFail<Record> testEquals
+        
+    [<Test; Category("FSharp type tests")>]
+    member __.``7. FSharp: struct record`` () = 
+        Check.QuickThrowOnFail<StructRecord> testEquals
 
     [<Test; Category("FSharp type tests")>]
     member __.``7. FSharp: cyclic record`` () =